Login Page

+++ project/app/views.py
@@ -1,8 +1,12 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render, redirect
 from django.http import HttpResponse
-
+from django.contrib.auth.models import User
+from django.contrib import messages   
 from .models import Students
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 
+@login_required(login_url='/login/')
 def hello(request):
     if(request.method == 'POST'):
         data = request.POST
@@ -28,11 +32,13 @@ def hello(request):
     con = {'hell':stu}
     return render(request,"j.html",con)
 
+@login_required(login_url='/login/')
 def dell(request, id):
-    print(id)
-    Students.objects.get(id=id).delete()
+    if request.method == "POST":
+        Students.objects.get(id=id).delete()
     return redirect('/')
 
+@login_required(login_url='/login/')
 def upda(request,id):
     record=Students.objects.all().get(id=id)
     if(request.method == 'POST'):
@@ -52,4 +58,51 @@ def upda(request,id):
     return render(request,'update.html',{'hell' : record})
 
 def register_pg(req):
-    return render(req,"register.html")
+    if req.method == 'POST':
+        data = req.POST
+        
+        fname = data.get('first_name')
+        lname = data.get('last_name')
+        username = data.get('username')
+        password = data.get('password')
+
+        if User.objects.filter(username=username).exists():
+            messages.error(req, "Username already exists")
+            return redirect('register')   
+
+        user = User.objects.create(
+            first_name=fname,
+            last_name=lname,
+            username=username
+        )
+        
+        user.set_password(password)   
+        user.save()
+
+        return redirect('/admin/')   
+
+    return render(req, "register.html")
+
+def login_pg(req):
+    if (req.method == "POST"):
+        data =req.POST
+        username = data.get("username")
+        password = data.get("password")
+
+        user=User.objects.filter(username = username)
+        if not user.exists():
+            messages.error(req,"Username is invalid")
+            return redirect('login')
+
+        user=authenticate(req,username=username,password=password)
+        if user is not None:
+            login(req,user)
+            return redirect("/") 
+        else:
+            messages.error(req,"Password is invalid")
+            return redirect("/login/")
+    return render(req, "login.html")
+
+def logout_pg(req):
+    logout(req)
+    return redirect("/login/")
